requirements/base.txt
BeautifulSoup==3.2.1
Django==1.6.5
Django-Select2==4.2.2
PIL==1.1.7
Pillow==2.3.0
South==0.8.4
amqp==1.0.6
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.19
distribute==0.6.34
-e git://github.com/tarvitz/django-actions.git@ab4a61cbf087f2f2a3e184b8c8fe93189b529873#egg=django_actions-dev
django-common-helpers==0.6.0
django-cron==0.3.1
django-debug-toolbar==1.2.1
django-debug-toolbar-extra==0.0.2
django-extensions==1.2.5
-e git://github.com/tarvitz/django-extwidgets.git@00141a36826c80c83162fc7284d9f8d655da4b47#egg=django_extwidgets-dev
django-filter==0.7
django-grappelli==2.5.0
django-picklefield==0.3.0
-e git+https://github.com/tarvitz/django-plop.git@49b7e9f14c48a27ab26aea4fddbf98018b0fb895#egg=django_plop-dev
django-recaptcha==1.0
django-taggit==0.10
django-tastypie==0.10.0
djangorestframework==3.1.1
gunicorn==17.5
ipdb==0.7
ipython==0.13.1
kombu==2.5.4
mimeparse==0.1.3
plop==0.1.1
psycopg2==2.5.1
python-dateutil==1.5
python-memcached==1.53
pytils==0.2.3
simplejson==2.6.2
six==1.2.0
sleekxmpp==1.1.11
sorl-thumbnail==11.12
sqlparse==0.1.11
tornado==3.1.1
wsgiref==0.1.2
